+++ src/api/v2/order.rs
@@ -291,6 +291,45 @@ impl Serialize for StopLoss {
 }
 
 
+/// An abstraction to be able to handle orders in both notional and quantity units.
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(untagged)]
+pub enum Amount {
+  /// Wrapper for the quantity field.
+  Quantity {
+    /// A number of shares to order. This can be a fractional number if
+    /// trading fractionals or a whole number if not.
+    #[serde(rename = "qty")]
+    quantity: Num,
+  },
+  /// Wrapper for the notional field.
+  Notional {
+    /// A dollar amount to use for the order. This can result in
+    /// fractional quantities.
+    #[serde(rename = "notional")]
+    notional: Num,
+  },
+}
+
+impl Amount {
+  /// Helper method to initialize a quantity.
+  #[inline]
+  pub fn quantity(amount: impl Into<Num>) -> Self {
+    Self::Quantity {
+      quantity: amount.into(),
+    }
+  }
+
+  /// Helper method to initialize a notional.
+  #[inline]
+  pub fn notional(amount: impl Into<Num>) -> Self {
+    Self::Notional {
+      notional: amount.into(),
+    }
+  }
+}
+
+
 /// A helper for initializing `OrderReq` objects.
 #[derive(Clone, Debug, Default, PartialEq)]
 pub struct OrderReqInit {
@@ -322,13 +361,13 @@ pub struct OrderReqInit {
 
 impl OrderReqInit {
   /// Create an `OrderReq` from an `OrderReqInit`.
-  pub fn init<S>(self, symbol: S, side: Side, quantity: u64) -> OrderReq
+  pub fn init<S>(self, symbol: S, side: Side, amount: Amount) -> OrderReq
   where
     S: Into<asset::Symbol>,
   {
     OrderReq {
       symbol: symbol.into(),
-      quantity,
+      amount,
       side,
       class: self.class,
       type_: self.type_,
@@ -352,9 +391,9 @@ pub struct OrderReq {
   /// Symbol or asset ID to identify the asset to trade.
   #[serde(rename = "symbol")]
   pub symbol: asset::Symbol,
-  /// Number of shares to trade.
-  #[serde(rename = "qty", serialize_with = "u64_to_str")]
-  pub quantity: u64,
+  /// Amount of shares to trade.
+  #[serde(flatten)]
+  pub amount: Amount,
   /// The side the order is on.
   #[serde(rename = "side")]
   pub side: Side,
@@ -507,13 +546,9 @@ pub struct Order {
   /// The symbol of the asset being traded.
   #[serde(rename = "symbol")]
   pub symbol: String,
-  /// The quantity being requested.
-  #[serde(
-    rename = "qty",
-    deserialize_with = "u64_from_str",
-    serialize_with = "u64_to_str"
-  )]
-  pub quantity: u64,
+  /// The amount being requested.
+  #[serde(flatten)]
+  pub amount: Amount,
   /// The quantity that was filled.
   #[serde(
     rename = "filled_qty",
@@ -715,6 +750,8 @@ EndpointNoParse! {
 mod tests {
   use super::*;
 
+  use std::str::FromStr as _;
+
   use futures::TryFutureExt;
 
   use serde_json::from_str as from_json;
@@ -766,6 +803,26 @@ mod tests {
     assert_eq!(to_json(&stop_loss).unwrap(), &expected[..]);
   }
 
+  /// Check that we can parse the `Amount::quantity` variant properly.
+  #[test]
+  fn parse_quantity_amount() {
+    let serialized = r#"{
+    "qty": "15"
+}"#;
+    let amount = from_json::<Amount>(&serialized).unwrap();
+    assert_eq!(amount, Amount::quantity(15));
+  }
+
+  /// Check that we can parse the `Amount::notional` variant properly.
+  #[test]
+  fn parse_notional_amount() {
+    let serialized = r#"{
+    "notional": "15.12"
+}"#;
+    let amount = from_json::<Amount>(&serialized).unwrap();
+    assert_eq!(amount, Amount::notional(Num::from_str("15.12").unwrap()));
+  }
+
   #[test]
   fn parse_reference_order() {
     let response = r#"{
@@ -802,7 +859,7 @@ mod tests {
       DateTime::parse_from_rfc3339("2018-10-05T05:48:59Z").unwrap()
     );
     assert_eq!(order.symbol, "AAPL");
-    assert_eq!(order.quantity, 15);
+    assert_eq!(order.amount, Amount::quantity(15));
     assert_eq!(order.type_, Type::Market);
     assert_eq!(order.time_in_force, TimeInForce::Day);
     assert_eq!(order.limit_price, Some(Num::from(107)));
@@ -820,7 +877,7 @@ mod tests {
         extended_hours,
         ..Default::default()
       }
-      .init(symbol, Side::Buy, 1);
+      .init(symbol, Side::Buy, Amount::quantity(1));
 
       let api_info = ApiInfo::from_env().unwrap();
       let client = Client::new(api_info);
@@ -829,7 +886,7 @@ mod tests {
       client.issue::<Delete>(&order.id).await.unwrap();
 
       assert_eq!(order.symbol, "SPY");
-      assert_eq!(order.quantity, 1);
+      assert_eq!(order.amount, Amount::quantity(1));
       assert_eq!(order.side, Side::Buy);
       assert_eq!(order.type_, Type::Limit);
       assert_eq!(order.time_in_force, TimeInForce::Day);
@@ -863,7 +920,7 @@ mod tests {
       trail_price: Some(Num::from(50)),
       ..Default::default()
     }
-    .init(symbol, Side::Buy, 1);
+    .init(symbol, Side::Buy, Amount::quantity(1));
 
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);
@@ -872,7 +929,7 @@ mod tests {
     client.issue::<Delete>(&order.id).await.unwrap();
 
     assert_eq!(order.symbol, "SPY");
-    assert_eq!(order.quantity, 1);
+    assert_eq!(order.amount, Amount::quantity(1));
     assert_eq!(order.side, Side::Buy);
     assert_eq!(order.type_, Type::TrailingStop);
     assert_eq!(order.time_in_force, TimeInForce::Day);
@@ -892,7 +949,7 @@ mod tests {
       trail_percent: Some(Num::from(10)),
       ..Default::default()
     }
-    .init(symbol, Side::Buy, 1);
+    .init(symbol, Side::Buy, Amount::quantity(1));
 
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);
@@ -901,7 +958,7 @@ mod tests {
     client.issue::<Delete>(&order.id).await.unwrap();
 
     assert_eq!(order.symbol, "SPY");
-    assert_eq!(order.quantity, 1);
+    assert_eq!(order.amount, Amount::quantity(1));
     assert_eq!(order.side, Side::Buy);
     assert_eq!(order.type_, Type::TrailingStop);
     assert_eq!(order.time_in_force, TimeInForce::Day);
@@ -922,7 +979,7 @@ mod tests {
       stop_loss: Some(StopLoss::Stop(Num::from(1))),
       ..Default::default()
     }
-    .init("SPY", Side::Buy, 1);
+    .init("SPY", Side::Buy, Amount::quantity(1));
 
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);
@@ -935,7 +992,7 @@ mod tests {
     }
 
     assert_eq!(order.symbol, "SPY");
-    assert_eq!(order.quantity, 1);
+    assert_eq!(order.amount, Amount::quantity(1));
     assert_eq!(order.side, Side::Buy);
     assert_eq!(order.type_, Type::Limit);
     assert_eq!(order.time_in_force, TimeInForce::Day);
@@ -956,7 +1013,7 @@ mod tests {
       stop_loss: Some(StopLoss::Stop(Num::from(1))),
       ..Default::default()
     }
-    .init("SPY", Side::Buy, 1);
+    .init("SPY", Side::Buy, Amount::quantity(1));
 
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);
@@ -969,7 +1026,7 @@ mod tests {
     }
 
     assert_eq!(order.symbol, "SPY");
-    assert_eq!(order.quantity, 1);
+    assert_eq!(order.amount, Amount::quantity(1));
     assert_eq!(order.side, Side::Buy);
     assert_eq!(order.type_, Type::Limit);
     assert_eq!(order.time_in_force, TimeInForce::Day);
@@ -993,7 +1050,7 @@ mod tests {
         limit_price: Some(Num::from(1)),
         ..Default::default()
       }
-      .init("AAPL", Side::Buy, 1);
+      .init("AAPL", Side::Buy, Amount::quantity(1));
 
       match client.issue::<Post>(&request).await {
         Ok(order) => {
@@ -1022,7 +1079,43 @@ mod tests {
       limit_price: Some(Num::from(1000)),
       ..Default::default()
     }
-    .init("AAPL", Side::Buy, 100000);
+    .init("AAPL", Side::Buy, Amount::quantity(100_000));
+
+    let result = client.issue::<Post>(&request).await;
+    let err = result.unwrap_err();
+
+    match err {
+      RequestError::Endpoint(PostError::InsufficientFunds(..)) => (),
+      _ => panic!("Received unexpected error: {:?}", err),
+    };
+  }
+
+  /// Test that we can submit an order with a notional amount.
+  #[test(tokio::test)]
+  async fn submit_unsatisfiable_notional_order() {
+    let request =
+      OrderReqInit::default().init("SPY", Side::Buy, Amount::notional(Num::new(10_000_000, 3)));
+
+    let api_info = ApiInfo::from_env().unwrap();
+    let client = Client::new(api_info);
+
+    let result = client.issue::<Post>(&request).await;
+    let err = result.unwrap_err();
+
+    match err {
+      RequestError::Endpoint(PostError::InsufficientFunds(..)) => (),
+      _ => panic!("Received unexpected error: {:?}", err),
+    };
+  }
+
+  /// Test that we can submit an order with a fractional quantity.
+  #[test(tokio::test)]
+  async fn submit_unsatisfiable_fractional_order() {
+    let qty = Num::from(1_000_000) + Num::new(1, 2);
+    let request = OrderReqInit::default().init("SPY", Side::Buy, Amount::quantity(qty));
+
+    let api_info = ApiInfo::from_env().unwrap();
+    let client = Client::new(api_info);
 
     let result = client.issue::<Post>(&request).await;
     let err = result.unwrap_err();
@@ -1062,7 +1155,7 @@ mod tests {
     assert_eq!(posted.asset_class, gotten.asset_class);
     assert_eq!(posted.asset_id, gotten.asset_id);
     assert_eq!(posted.symbol, gotten.symbol);
-    assert_eq!(posted.quantity, gotten.quantity);
+    assert_eq!(posted.amount, gotten.amount);
     assert_eq!(posted.type_, gotten.type_);
     assert_eq!(posted.side, gotten.side);
     assert_eq!(posted.time_in_force, gotten.time_in_force);
@@ -1088,7 +1181,7 @@ mod tests {
       extended_hours: true,
       ..Default::default()
     }
-    .init("SPY", Side::Buy, 1);
+    .init("SPY", Side::Buy, Amount::quantity(1));
 
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);
@@ -1111,7 +1204,7 @@ mod tests {
       limit_price: Some(Num::from(1)),
       ..Default::default()
     }
-    .init("AAPL", Side::Buy, 1);
+    .init("AAPL", Side::Buy, Amount::quantity(1));
 
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);
@@ -1136,7 +1229,7 @@ mod tests {
 
     match result {
       Ok(order) => {
-        assert_eq!(order.quantity, 2);
+        assert_eq!(order.amount, Amount::quantity(2));
         assert_eq!(order.time_in_force, TimeInForce::UntilCanceled);
         assert_eq!(order.limit_price, Some(Num::from(2)));
         assert_eq!(order.stop_price, None);
@@ -1159,7 +1252,7 @@ mod tests {
       trail_price: Some(Num::from(20)),
       ..Default::default()
     }
-    .init("SPY", Side::Buy, 1);
+    .init("SPY", Side::Buy, Amount::quantity(1));
 
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);
@@ -1203,7 +1296,7 @@ mod tests {
       client_order_id: Some(client_order_id.clone()),
       ..Default::default()
     }
-    .init("SPY", Side::Buy, 1);
+    .init("SPY", Side::Buy, Amount::quantity(1));
 
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);